\name{numbers-package}
\alias{numbers-package}
\alias{numbers}
\docType{package}
\title{
\packageTitle{numbers}
}
\description{
\packageDescription{numbers}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{numbers}
\packageIndices{numbers}

Although R does not have a true integer data type, integers can be
represented exactly up to 2^53-1 . The numbers package attempts to 
provided basic number-theoretic functions that will work correcty
and relatively fast up to this level.
}
\author{
\packageAuthor{numbers}

Maintainer: \packageMaintainer{numbers}
}
\references{
  Hardy, G. H., and E. M. Wright (1980). An Introduction to the
  Theory of Numbers. 5th Edition, Oxford University Press.

  Riesel, H. (1994). Prime Numbers and Computer Methods for
  Factorization. Second Edition, Birkhaeuser Boston.

  Crandall, R., and C. Pomerance (2005). Prime Numbers:
  A Computational Perspective. Springer Science+Business.

  Shoup, V. (2009). A Computational Introduction to Number Theory
  and Algebra. Second Edition, Cambridge University Press.

  Arndt, J. (2010). Matters Computational: Ideas, Algorithms,
  Source Code. 2011 Edition, Springer-Verlag, Berlin Heidelberg.

  Forster, O. (2014). Algorithmische Zahlentheorie. 2. Auflage,
  Springer Spektrum Wiesbaden.
}
