\name{numbers-package}
\alias{numbers-package}
\docType{package}
\title{
  Number-Theoretic Functions for R
}
\description{
  Number-theoretic Functions for R
}
\details{
\tabular{ll}{
Package: \tab numbers\cr
Type: \tab Package\cr
Version: \tab 0.4.3\cr
Date: \tab 2013-11-16\cr
License: \tab GPL (>= 3)\cr
}
}
\author{
Hans Werner Borchers

Maintainer: Hans W. Borchers \email{hwborchers@googlemail.com}
}
\references{
  Hardy and Wright. The Theory of Numbers
}
